+++ src/test/multisig_tests.cpp
@@ -5,7 +5,7 @@
 #include "consensus/upgrades.h"
 #include "key.h"
 #include "keystore.h"
-#include "main.h"
+#include "policy/policy.h"
 #include "script/script.h"
 #include "script/script_error.h"
 #include "script/interpreter.h"
